server/settings/vars/common_vars.py
import pathlib

import decouple

# Build paths inside the project like this: BASE_DIR.joinpath('some')
# `pathlib` is better than writing: dirname(dirname(dirname(__file__)))
BASE_DIR = pathlib.PurePath(__file__).parent.parent.parent.parent

# Loading `.env` files
# See docs: https://gitlab.com/mkleehammer/autoconfig
config = decouple.AutoConfig(search_path=BASE_DIR.joinpath('config'))
